+++ src/app/global-variables.ts
@@ -1,12 +1,12 @@
 
-import { animate, animateChild, group, query, style, transition, trigger } from '@angular/animations';
+import { animate, animateChild, group, query, state, style, transition, trigger } from '@angular/animations';
 import { LinkInfo } from "./classes/LinkInfo";
 import { InvestmentPageComponent } from './pages/investment-page/investment-page.component';
 import { InvestmentVersusPageComponent } from './pages/investment-versus-page/investment-versus-page.component';
 import { PageProjectsAndExamplesComponent } from './pages/page-projects-and-examples/page-projects-and-examples.component';
 
 export const LINK_INFOS: LinkInfo[] = [
-  new LinkInfo("Investment", 'investment/versus', InvestmentVersusPageComponent, [
+  new LinkInfo("Investment", 'investment', InvestmentVersusPageComponent, [
     new LinkInfo("DCA versus VA", 'investment/versus', InvestmentVersusPageComponent), //Dollar Cost Averaging, Value Averaging
     new LinkInfo("Custom", 'investment/custom', InvestmentPageComponent),
   ]),
@@ -38,5 +38,24 @@ export const FADE_ANIMATION =
         ], { optional: true })
       ]),
       query(':enter', animateChild(), { optional: true }),
+    ]),
+  ]);
+
+export const MENU_BTN_ANIMATION =
+  trigger(
+    'menuBtnAnimation', [
+    state('show', style({
+      width: '40px',
+      opacity: 1,
+    })),
+    state('hidden', style({
+      width: '0px',
+      opacity: 0,
+    })),
+    transition('show => hidden', [
+      animate('0.2s')
+    ]),
+    transition('hidden => show', [
+      animate('0.2s')
     ]),
   ]);