+++ ChangeLog.md
@@ -5,6 +5,8 @@
 
 1. Add route fade animation
 2. Change global-variables and add global-methods
+3. Change LINK_INFOS
+4. remove toolbar component
 
 2021/09/09